simpilified classes into smaller modules
