--- kwhelp/error/__init__.py
@@ -1,3 +1,0 @@
-# coding: utf-8
-class CancelEventError(Exception):
-    '''Cancel Event Error'''
