--- kwhelp/helper/builder.py
@@ -1,133 +1,0 @@
-# coding: utf-8
-from collections import UserList
-from typing import List, Optional, Callable
-from .. rules import IRule
-from .. helper.args import HelperArgs
-
-# region class AssignBuilder
-class AssignBuilder(UserList):
-    '''Helper class for building list to use with "KwargsHelper.Assing() method'''
-
-    def __init__(self) -> None:
-        super().__init__(initlist=None)
-        self._keys = set()
-
-    def append(self, key: str, field: Optional[str] = None, require: bool = False, default: Optional[object] = None, types: Optional[List[str]] = None, rules: Optional[List[Callable[[IRule], bool]]] = None):
-        '''
-        Appends dictionary item of parameters to list
-        @key: Type:str, the key of the key, value pair.
-        @field: (optional) Type:str the name of the field.
-        @require: (optional) Type:bool Default: `False`
-        @default: (optional) Type:str, default value to assign.
-        @types: (optional) Type:List[str], a string list of one or more types that the value of the key value pair must match.
-        @rules: (optional) Type:List[Callable[[IRule], bool]]
-        '''
-        if not isinstance(key, str):
-            raise TypeError(self._get_type_error_method_msg(
-                method_name='append', arg=key, arg_name='key', expected_type='str'
-            ))
-        _key = key.strip()
-        if len(_key) == 0:
-            raise ValueError(self._get_value_error_msg(
-                method_name='append', arg=key, arg_name='key',
-                msg="can not be empty or whitespace"
-            ))
-        if _key in self._keys:
-            raise ValueError(
-                self._get_value_error_msg(
-                    method_name='append', arg=key, arg_name='key',
-                    msg='already exist.'
-                )
-            )
-        _args = HelperArgs(key=_key)
-        _args.require = require
-        if field is not None:
-            _args.field = field
-        if default is not None:
-            _args.default = default
-        if types is not None:
-            _args.types = set(types)
-        if rules is not None:
-            _args.rules = rules
-
-        super().append(_args)
-        self._keys.add(_key)
-        return None
-
-    def append_helper(self, helper: HelperArgs):
-        '''
-        Appends dictionary item of parameters to list
-        @helper: args to add
-        '''
-        if not isinstance(helper, HelperArgs):
-            raise TypeError(self._get_type_error_method_msg(
-                method_name='append_helper', arg=helper, arg_name='helper', expected_type='HelperArgs'
-            ))
-        if len(helper.key) == 0:
-            raise ValueError(self._get_value_error_msg(
-                method_name='append_helper', arg=helper, arg_name='helper.key',
-                msg="can not be empty or whitespace"
-            ))
-
-        if helper.key in self._keys:
-            raise ValueError(
-                self._get_value_error_msg(
-                    method_name='append_helper', arg=helper, arg_name='key',
-                    msg='already exist.'
-                )
-            )
-        super().append(helper)
-        self._keys.add(helper.key)
-
-    def remove(self, item: HelperArgs) -> None:
-        if item is None:
-            return None
-        if not isinstance(item, HelperArgs):
-            raise TypeError(self._get_type_error_method_msg(
-                method_name='remove', arg=item, arg_name='item', expected_type='HelperArgs'
-            ))
-        _key = item.key
-        super().remove(item)
-        self._keys.remove(_key)
-        return None
-
-    def extend(self, other: 'AssignBuilder') -> None:
-        if not isinstance(other, AssignBuilder):
-            raise NotImplementedError(
-                f"{self.__class__.__name__}.extend() only supports extending by instances of 'AssignBuilder'")
-        for item in other:
-            key: str = item.key
-            if not key in self._keys:
-                super().append(item)
-                self._keys.add(key)
-
-    def _get_type_error_method_msg(self, method_name: str, arg: object, arg_name: str, expected_type: str) -> str:
-        result = f"{self.__class__.__name__}.{method_name}() arg '{arg_name}' is expecting type of '{expected_type}'. Got type of '{type(arg).__name__}'"
-        return result
-
-    def _get_value_error_msg(self, method_name: str, arg: object, arg_name: str, msg: str) -> str:
-        result = f"{self.__class__.__name__}.{method_name}() arg '{arg_name}' {msg}"
-        return result
-
-    # region dunder methods
-    def __getitem__(self, i: int) -> HelperArgs:
-        return super().__getitem__(i)
-
-    def __setitem__(self, i: int, helper: HelperArgs):
-        if not isinstance(helper, HelperArgs):
-            raise TypeError(self._get_type_error_method_msg(
-                method_name='__setitem__', arg=helper, arg_name='helper', expected_type='HelperArgs'
-            ))
-        if i < 0 or i >= len(self):
-            raise IndexError
-        current = self[i]
-        if helper.key != current.key:
-            if helper.key in self._keys:
-                raise ValueError(
-                    self._get_value_error_msg(
-                        method_name='__setitem__', arg=helper, arg_name='key',
-                        msg='already exist.'
-                    ))
-        super().__setitem__(i, helper)
-    # endregion dunder methods
-# endregion class AssignBuilder
