+++ tests/run_pytest.py
@@ -2,7 +2,7 @@
 from subprocess import run
 
 def main():
-    cmd_str = 'pytest --cov=src.main --cov=src.error --cov=src.event_args --cov=src.helper --cov=src.rules --cov-report=html'
+    cmd_str = 'pytest --cov=kwhelp --cov=kwhelp.rules --cov=kwhelp.helper --cov-report=html'
     res = run(cmd_str.split(' '))
     print(res.stdout)
